updated small part of the code

+++ rand_ex.m
@@ -1,6 +1,6 @@
 /*Produce a random number in the range [a,b]*/
 double rand_draw(double a, double b) {
-  double random = ((float) rand()) / (float) RAND_MAX;
+  double random = ((double) rand()) / (double) RAND_MAX;
   double diff = b - a;
   double r = random * diff;
   return a + r;
